unified-proxy-collector/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "unified-proxy-collector"
version = "1.0.0"
description = "A unified high-performance proxy collector merging features from V2Ray-Collector, TG-Parser, and Telegram-Configs-Collector."
readme = "README.md"
requires-python = ">=3.8"
authors = [
  { name = "Unified Proxy Collector Team" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "beautifulsoup4>=4.14.3",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "wget>=3.2",
    "jdatetime>=5.2.0",
    "tldextract>=5.3.0",
    "geoip2>=5.2.0",
    "dnspython>=2.8.0",
    "pycountry_convert>=0.7.2",
    "rich>=14.2.0",
    "colorama>=0.4.6",
    "textual>=6.7.1",
    "pyyaml"
]

[project.scripts]
unified-collector = "unified_proxy_collector.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["unified_proxy_collector*"]
